# backend/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go microservice API for Quality Studio Server.",
        "title": "Quality Studio API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/konflux-ci/quality-studio"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/konflux-ci/quality-studio/blob/main/LICENSE"
        },
        "version": "v1alpha"
    },
    "host": "127.0.0.1:9898",
    "basePath": "/api/quality/",
    "paths": {
...
        "/prow/results/get": {
            "get": {
                "description": "returns all prow jobs related to git_organization and repository_name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prow Jobs info"
                ],
                "summary": "Prow Jobs info",
                "parameters": [
                    {
                        "description": "repository name",
                        "name": "repository",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prow.GitRepositoryRequest"
                        }
                    },
                    {
                        "description": "git_organization",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prow.GitRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/prow/results/latest/get": {
            "get": {
                "description": "returns all prow jobs related to git_organization and repository_name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prow Jobs info"
                ],
                "summary": "Prow Jobs info",
                "parameters": [
                    {
                        "description": "repository name",
                        "name": "repository",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prow.GitRepositoryRequest"
                        }
                    },
                    {
                        "description": "git_organization",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prow.GitRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "prow.GitRepositoryRequest": {
            "type": "object",
            "properties": {
                "git_organization": {
                    "type": "string"
                },
                "repository_name": {
                    "type": "string"
                }
            }
        },
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "The error message.\nRequired: true",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The error message.\nRequired: false",
                    "type": "integer"
                }
            }
        },
...
    }
}